# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Parallax API",
    "description": "Backend API for manually submitted Parallax missions. External writes are proposed, approved, executed, verified, and audited by deterministic backend services.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/missions": {
      "post": {
        "tags": [
          "missions"
        ],
        "summary": "Submit a manual mission",
        "description": "Persist a manual mission and its outbox job atomically; no agent runs inline.",
        "operationId": "create_mission_api_missions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "X-Workspace-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MissionCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionRead"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Request validation failed"
          }
        }
      },
      "get": {
        "tags": [
          "missions"
        ],
        "summary": "List missions",
        "operationId": "list_missions_api_missions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "X-Workspace-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MissionRead"
                  },
                  "title": "Response List Missions Api Missions Get"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Request validation failed"
          }
        }
      }
    },
    "/api/missions/{mission_id}": {
      "get": {
        "tags": [
          "missions"
        ],
        "summary": "Get a mission",
        "operationId": "get_mission_api_missions__mission_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Mission Id"
            }
          },
          {
            "name": "X-Workspace-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionRead"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Request validation failed"
          }
        }
      }
    },
    "/api/missions/{mission_id}/context-pack": {
      "get": {
        "tags": [
          "missions"
        ],
        "summary": "Get a mission Context Pack",
        "operationId": "get_context_pack_api_missions__mission_id__context_pack_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Mission Id"
            }
          },
          {
            "name": "X-Workspace-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextPackRead"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Request validation failed"
          }
        }
      }
    },
    "/api/missions/{mission_id}/assessment": {
      "get": {
        "tags": [
          "missions"
        ],
        "summary": "Get a validated Agent assessment",
        "operationId": "get_assessment_api_missions__mission_id__assessment_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Mission Id"
            }
          },
          {
            "name": "X-Workspace-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentAssessmentRead"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Request validation failed"
          }
        }
      }
    },
    "/api/missions/{mission_id}/cancel": {
      "post": {
        "tags": [
          "missions"
        ],
        "summary": "Cancel a mission",
        "operationId": "cancel_mission_api_missions__mission_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Mission Id"
            }
          },
          {
            "name": "X-Workspace-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionRead"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Request validation failed"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentAssessmentRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "mission_id": {
            "type": "string",
            "format": "uuid",
            "title": "Mission Id"
          },
          "mode": {
            "type": "string",
            "enum": [
              "service",
              "fallback"
            ],
            "title": "Mode"
          },
          "context_summary": {
            "type": "string",
            "title": "Context Summary"
          },
          "risk_level": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ],
            "title": "Risk Level"
          },
          "risk_factors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Risk Factors"
          },
          "review_effort_minutes": {
            "type": "integer",
            "title": "Review Effort Minutes"
          },
          "effort_rationale": {
            "type": "string",
            "title": "Effort Rationale"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "reviewer_candidates": {
            "items": {
              "$ref": "#/components/schemas/ReviewerCandidateRead"
            },
            "type": "array",
            "title": "Reviewer Candidates"
          },
          "citations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Citations"
          },
          "proposals": {
            "items": {
              "$ref": "#/components/schemas/AgentProposalRead"
            },
            "type": "array",
            "title": "Proposals"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "mission_id",
          "mode",
          "context_summary",
          "risk_level",
          "risk_factors",
          "review_effort_minutes",
          "effort_rationale",
          "confidence",
          "explanation",
          "reviewer_candidates",
          "citations",
          "proposals",
          "created_at"
        ],
        "title": "AgentAssessmentRead"
      },
...
      "ContextPackRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "mission_id": {
            "type": "string",
            "format": "uuid",
            "title": "Mission Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "content": {
            "type": "object",
            "title": "Content"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceRead"
            },
            "type": "array",
            "title": "Evidence"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "mission_id",
          "version",
          "summary",
          "content",
          "content_hash",
          "evidence",
          "created_at"
        ],
        "title": "ContextPackRead"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
...
      "MissionCreate": {
        "properties": {
          "prompt": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 3,
            "title": "Prompt"
          },
          "project": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Project",
            "default": "General"
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "MissionCreate"
      },
      "MissionRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "project": {
            "type": "string",
            "title": "Project"
          },
          "status": {
            "$ref": "#/components/schemas/MissionStatus"
          },
          "progress_current": {
            "type": "integer",
            "title": "Progress Current"
          },
          "progress_total": {
            "type": "integer",
            "title": "Progress Total"
          },
          "result_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Summary"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/MissionStepRead"
            },
            "type": "array",
            "title": "Steps"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "prompt",
          "project",
          "status",
          "progress_current",
          "progress_total",
          "result_summary",
          "steps",
          "created_at",
          "updated_at"
        ],
        "title": "MissionRead"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}